dymanic/requirements.txt
appdirs==1.4.4
asgiref==3.5.2
cached-property==1.5.2
chardet==3.0.4
distlib==0.3.4
Django==4.1.1
djangorestframework==3.13.1
filelock==3.7.1
lxml==4.9.1
num2words==0.5.6
passlib==1.7.3
platformdirs==2.5.2
polib==1.1.0
pycparser==2.21
pyparsing==3.0.9
PyPDF2==1.26.0
pypiwin32==223
pyserial==3.4
python-dateutil==2.7.3
python-stdnum==1.13
pytz==2019.3
pyusb==1.0.2
pywin32==304
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
suds==1.1.2
tzdata==2022.2
urllib3==1.26.5
virtualenv==20.15.0
Werkzeug==2.0.2
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.4.0
